# Calhacks/src/components/api/schemas/ImgRecogModel/models/Error.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Error = {
    /**
     * Error message describing what went wrong
     */
    error: string;
    /**
     * Additional error details or context
     */
    message?: string;
};